notice/requirements.txt
django==3.2.5
channels
celery
django-celery-beat
djangorestframework
mysqlclient

